@@ -23,15 +23,15 @@
   "dependencies": {
     "@instructure/config-loader": "^5",
     "@instructure/ui-docs-client": "^5",
-    "favicons-webpack-plugin": "^0.0.9",
-    "globby": "^8.0.2",
-    "gray-matter": "^4.0.1",
-    "html-webpack-plugin": "^2.29.0",
+    "favicons-webpack-plugin": "junyper/favicons-webpack-plugin#master",
+    "globby": "^9.0.0",
+    "gray-matter": "3.1.1",
+    "html-webpack-plugin": "^3.2.0",
     "jsdoc-api": "^4.0.1",
     "loader-utils": "^1.2.3",
     "react-docgen": "3.0.0-rc.1",
-    "script-ext-html-webpack-plugin": "^1.8.8",
-    "webpack": "^3"
+    "script-ext-html-webpack-plugin": "^2.1.3",
+    "webpack": "^4.28.1"
   },
   "publishConfig": {
     "access": "public"
