@@ -16,14 +16,14 @@
   },
   "license": "MIT",
   "dependencies": {
-    "babel-eslint": "^8.2.5",
-    "eslint": "^5.7.0",
-    "eslint-plugin-import": "^2.14.0",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^5.12.1",
+    "eslint-plugin-import": "^2.15.0",
     "eslint-plugin-instructure-ui": "^5",
     "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-mocha": "^5.2.1",
     "eslint-plugin-notice": "^0.7.7",
-    "eslint-plugin-react": "^7.12.3"
+    "eslint-plugin-react": "^7.12.4"
   },
   "publishConfig": {
     "access": "public"
