@@ -21,30 +21,31 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@babel/cli": "^7.2.3",
     "@instructure/cz-lerna-changelog": "^5",
     "@instructure/pkg-utils": "^5",
-    "babel-cli": "^6.26.0",
     "chalk": "^2.4.2",
     "commitizen": "^3.0.5",
     "concurrently": "^4.1.0",
     "cross-env": "^5.2.0",
     "cross-spawn": "^6.0.4",
     "dotenv": "^6.2.0",
-    "eslint": "^5.12.0",
+    "eslint": "^5.12.1",
     "gh-pages": "^2.0.1",
-    "jira-client": "^6.4.1",
-    "karma": "^3.1.4",
+    "jira-client": "^6.5.0",
+    "karma-cli": "^2.0.0",
     "lerna": "3.0.0-beta.21",
     "rimraf": "^2.6.3",
     "standard-version": "^4.4.0",
-    "stylelint": "^9.3.0",
+    "stylelint": "^9.10.1",
     "validate-commit-msg": "^2.14.0",
     "which": "^1.3.1",
-    "webpack": "^3",
-    "webpack-dev-server": "^2.11.0"
+    "webpack-cli": "^3.2.1",
+    "webpack-dev-server": "^3.1.14"
   },
   "peerDependencies": {
-    "react": "^0.14.9 || ^15 || ^16"
+    "react": "^0.14.9 || ^15 || ^16",
+    "webpack": "^4"
   },
   "publishConfig": {
     "access": "public"
