@@ -19,27 +19,19 @@
   "dependencies": {
     "@instructure/babel-plugin-themeable-styles": "^5",
     "@instructure/config-loader": "^5",
-    "@instructure/pkg-utils": "^5",
-    "@instructure/postcss-themeable-styles": "^5",
-    "@instructure/ui-babel-preset": "^5",
-    "@instructure/ui-postcss-config": "^5",
-    "babel-loader": "^7.1.5",
-    "css-loader": "^0.28.9",
+    "babel-loader": "^8.0.5",
+    "css-loader": "^2.1.0",
     "eslint-loader": "^2.1.1",
-    "file-loader": "^1.1.5",
-    "globby": "^8.0.1",
-    "happypack": "^5.0.1",
-    "loader-utils": "^1.2.3",
-    "postcss": "^6.0.23",
-    "postcss-browser-reporter": "^0.5.0",
-    "postcss-import": "11.1.0",
-    "postcss-modules": "1.2.0",
-    "postcss-reporter": "^6.0.0",
-    "react-docgen": "3.0.0-rc.1",
-    "style-loader": "^0.21.0",
-    "stylelint": "^9.3.0",
-    "url-loader": "^1.0.1",
-    "webpack": "^3"
+    "file-loader": "^3.0.1",
+    "postcss-loader": "^3",
+    "postcss-reporter": "^6.0.1",
+    "react-dev-utils": "^7.0.1",
+    "style-loader": "^0.23.1",
+    "stylelint": "^9.10.1",
+    "thread-loader": "^2.1.1",
+    "terser-webpack-plugin": "^1.2.1",
+    "url-loader": "^1.1.2",
+    "webpack": "^4"
   },
   "publishConfig": {
     "access": "public"
