@@ -23,20 +23,12 @@
  */
 
 const path = require('path')
-const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
 const DocsPlugin = require('@instructure/ui-docs-plugin')
-const webpack = require('webpack')
-
-const ENV = process.env.NODE_ENV
-const DEBUG = Boolean(process.env.DEBUG) || ENV === 'development'
-
 const projectRoot = path.resolve(__dirname, '../../')
 // eslint-disable-next-line instructure-ui/no-relative-package-imports
 const pkg = require('../../package.json')
 
-let plugins = require('@instructure/ui-webpack-config').plugins
-
-plugins = plugins.concat([
+module.exports = [
   new DocsPlugin({
     projectRoot,
     title: `${pkg.name} : ${pkg.description} (${pkg.version})`,
@@ -131,32 +123,5 @@ plugins = plugins.concat([
       }
     },
     template: './index.ejs'
-  }),
-  new webpack.optimize.CommonsChunkPlugin({
-    // should match entries in webpack.config.js
-    // the ui-docs entry is generated by the ui-docs-plugin
-    name: ['common', 'globals'].reverse(),
-    minChunks: Infinity
   })
-])
-
-if (!DEBUG) {
-  plugins = plugins.concat([
-    new UglifyJsPlugin({
-      sourceMap: true,
-      parallel: true,
-      uglifyOptions: {
-        ecma: 5,
-        mangle: false, // leaves variable names as-is so it is easier to debug, gzip takes care of most of this anyway
-        compress: {
-          sequences: false // prevents it from combining a bunch of statements with ","s so it is easier to set breakpoints
-        },
-        output: {
-          semicolons: false // prevents everything being on one line so it's easer to read in devtools
-        }
-      }
-    })
-  ])
-}
-
-module.exports = plugins
+]
