@@ -21,13 +21,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-
- module.exports = {
-   presets: [[
-     require('@instructure/ui-babel-preset'),
-     {
-      coverage: Boolean(process.env.COVERAGE),
-      esModules: Boolean(process.env.ES_MODULES)
-     }
-   ]]
- }
+module.exports =  {
+  presets: ['@instructure/ui-babel-preset']
+}
