@@ -22,9 +22,8 @@
  * SOFTWARE.
  */
 
-module.exports = {
-  presets: [[ require('@instructure/ui-babel-preset'), {
-    coverage: Boolean(process.env.COVERAGE),
-    esModules: Boolean(process.env.ES_MODULES)
-  }]]
-}
+ module.exports = {
+   presets: [[ require('@instructure/ui-babel-preset'), {
+     esModules: Boolean(process.env.ES_MODULES)
+   }]]
+ }
